test(fixtures): consolidate authored program corpus
