+++ tests/fixtures/logical/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable authored programs whose computation deliberately carries no placement."""
