+++ tests/fixtures/placed/__init__.py
@@ -1,0 +1,4 @@
+"""Reusable authored programs with at least one explicitly placed execution path.
+
+Programs that isolate syntax or one mechanism remain in their owning test module.
+"""
