+++ tests/installed/smoke_check.py
@@ -34,10 +34,12 @@ def test_check_help_explains_input_and_output_positions(tf) -> None:
     assert done.stderr == ""
 
 
-def test_a_twin_is_compared_against_the_module_it_states(tf, twin) -> None:
+def test_a_twin_is_compared_against_the_module_it_states(
+    tf, square_twin, fused_twin
+) -> None:
     faithful = tf(
         "check",
-        f"{twin}:Twin.main",
+        f"{square_twin}:Twin.main",
         *_ARGS,
         "--fn",
         "cosine",
@@ -48,23 +50,25 @@ def test_a_twin_is_compared_against_the_module_it_states(tf, twin) -> None:
     assert "reference: evaluator on Model.main" in faithful.stdout
     assert "max_violation 0" in faithful.stdout
 
-    drifted = tf("check", f"{twin}:Drifted.main", *_ARGS)
+    drifted = tf("check", f"{square_twin}:Drifted.main", *_ARGS)
     assert drifted.returncode == 1
     assert "FAIL" in drifted.stdout
 
-    fused = tf("check", f"{twin}:FusedTwin.fused", *_ARGS)
+    fused = tf("check", f"{fused_twin}:FusedTwin.fused", *_ARGS)
     assert fused.returncode == 0, fused.stderr
     assert "reference: evaluator on Fused.fused" in fused.stdout
 
 
-def test_a_whole_module_is_checked_against_an_expected_output_file(tf, twin, tmp_path) -> None:
+def test_a_whole_module_is_checked_against_an_expected_output_file(
+    tf, square_twin, tmp_path
+) -> None:
     activation = torch.arange(168, dtype=torch.float32)
     torch.save(activation, tmp_path / "x.pt")
     torch.save(activation * activation, tmp_path / "expected.pt")
 
     done = tf(
         "check",
-        f"{twin}:Model",
+        f"{square_twin}:Model",
         "--input",
         str(tmp_path / "x.pt"),
         "--expected",
@@ -80,7 +84,9 @@ def test_a_whole_module_is_checked_against_an_expected_output_file(tf, twin, tmp
     assert "elements 168" in done.stdout
 
 
-def test_a_nested_activation_file_supplies_one_orchestration_parameter(tf, twin, tmp_path) -> None:
+def test_a_nested_activation_file_supplies_one_orchestration_parameter(
+    tf, orchestrated_twin, tmp_path
+) -> None:
     hidden = torch.arange(168, dtype=torch.float32)
     mixer_args = (
         torch.ones(168),
@@ -97,7 +103,7 @@ def test_a_nested_activation_file_supplies_one_orchestration_parameter(tf, twin,
 
     compared = tf(
         "check",
-        f"{twin}:OrchestratedTwin",
+        f"{orchestrated_twin}:OrchestratedTwin",
         "--input",
         str(tmp_path / "hidden.pt"),
         "--input",
@@ -121,7 +127,7 @@ def test_a_nested_activation_file_supplies_one_orchestration_parameter(tf, twin,
 
     expected = tf(
         "check",
-        f"{twin}:Orchestrated",
+        f"{orchestrated_twin}:Orchestrated",
         "--input",
         str(tmp_path / "hidden.pt"),
         "--input",
@@ -140,7 +146,9 @@ def test_a_nested_activation_file_supplies_one_orchestration_parameter(tf, twin,
     assert expected.returncode == 0, expected.stderr
 
 
-def test_a_non_tensor_nested_activation_leaf_names_its_position(tf, twin, tmp_path) -> None:
+def test_a_non_tensor_nested_activation_leaf_names_its_position(
+    tf, orchestrated_twin, tmp_path
+) -> None:
     torch.save(torch.arange(168, dtype=torch.float32), tmp_path / "hidden.pt")
     torch.save(
         (torch.ones(168), ("not a tensor", torch.ones(168)), torch.ones(168), torch.ones(168)),
@@ -149,7 +157,7 @@ def test_a_non_tensor_nested_activation_leaf_names_its_position(tf, twin, tmp_pa
 
     done = tf(
         "check",
-        f"{twin}:OrchestratedTwin",
+        f"{orchestrated_twin}:OrchestratedTwin",
         "--input",
         str(tmp_path / "hidden.pt"),
         "--input",
@@ -167,10 +175,10 @@ def test_a_non_tensor_nested_activation_leaf_names_its_position(tf, twin, tmp_pa
     assert "mixer_args.pt[1][0]" in done.stderr
 
 
-def test_an_orchestration_method_names_the_files_its_inputs_need(tf, twin) -> None:
+def test_an_orchestration_method_names_the_files_its_inputs_need(tf, orchestrated_twin) -> None:
     done = tf(
         "check",
-        f"{twin}:OrchestratedTwin",
+        f"{orchestrated_twin}:OrchestratedTwin",
         "--inputs",
         "random",
         "--out",
@@ -186,13 +194,13 @@ def test_an_orchestration_method_names_the_files_its_inputs_need(tf, twin) -> No
     assert 'torch.save((...), "mixer_args.pt")' in done.stderr
 
 
-def test_the_inputs_are_exactly_one_form(tf, twin, tmp_path) -> None:
+def test_the_inputs_are_exactly_one_form(tf, square_twin, tmp_path) -> None:
     torch.save(torch.arange(168, dtype=torch.float32), tmp_path / "x.pt")
 
     for form in ("random", "real"):
         done = tf(
             "check",
-            f"{twin}:Twin.main",
+            f"{square_twin}:Twin.main",
             "--input",
             str(tmp_path / "x.pt"),
             "--inputs",
@@ -206,10 +214,10 @@ def test_the_inputs_are_exactly_one_form(tf, twin, tmp_path) -> None:
         assert "give exactly one form" in done.stderr
 
 
-def test_two_entirely_zero_sides_are_a_match_not_a_total_mismatch(tf, twin) -> None:
+def test_two_entirely_zero_sides_are_a_match_not_a_total_mismatch(tf, square_twin) -> None:
     done = tf(
         "check",
-        f"{twin}:Twin.zeroed",
+        f"{square_twin}:Twin.zeroed",
         "--inputs",
         "random",
         "--out",
@@ -232,13 +240,13 @@ def test_two_entirely_zero_sides_are_a_match_not_a_total_mismatch(tf, twin) -> N
 
 
 def test_real_weights_come_from_the_checkpoint_and_activations_are_drawn(
-    tf, twin, tmp_path
+    tf, weighted_twin, tmp_path
 ) -> None:
     save_file({"w": torch.linspace(0.5, 2.0, 168)}, str(tmp_path / "model.safetensors"))
 
     done = tf(
         "check",
-        f"{twin}:WeightedRootTwin.scaled",
+        f"{weighted_twin}:WeightedRootTwin.scaled",
         "--inputs",
         "real",
         "--ckpt",
@@ -260,27 +268,31 @@ def test_real_weights_come_from_the_checkpoint_and_activations_are_drawn(
     assert "weights the checkpoint actual torch.float32 (declared f32)" in done.stdout
     assert "max_violation 0" in done.stdout
 
-    refused = tf("check", f"{twin}:WeightedRootTwin.scaled", "--inputs", "real", *_ARGS[2:])
+    refused = tf(
+        "check", f"{weighted_twin}:WeightedRootTwin.scaled", "--inputs", "real", *_ARGS[2:]
+    )
     assert refused.returncode == 1
     assert "needs --ckpt DIR" in refused.stderr
 
 
-def test_a_nested_twin_is_reached_through_the_child_it_is_declared_under(tf, twin) -> None:
-    done = tf("check", f"{twin}:NestedTwin.child.scaled", *_ARGS)
+def test_a_nested_twin_is_reached_through_the_child_it_is_declared_under(tf, nested_twin) -> None:
+    done = tf("check", f"{nested_twin}:NestedTwin.child.scaled", *_ARGS)
     assert done.returncode == 0, done.stderr
 
     assert "reference: evaluator on child.scaled" in done.stdout
     assert "max_violation 0" in done.stdout
 
 
-def test_a_runtime_module_that_names_no_authored_module_is_refused(tf, twin) -> None:
-    for target, refused in (
-        ("Handwritten", "names no authored Module"),
-        ("Mislabelled", "module must be Module or None, got str"),
+def test_a_runtime_module_that_names_no_authored_module_is_refused(
+    tf, handwritten_twin, mislabelled_twin
+) -> None:
+    for source, target, refused in (
+        (handwritten_twin, "Handwritten", "names no authored Module"),
+        (mislabelled_twin, "Mislabelled", "module must be Module or None, got str"),
     ):
         done = tf(
             "check",
-            f"{twin}:{target}",
+            f"{source}:{target}",
             "--inputs",
             "random",
             "--out",
